+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 # Todo
+- jslint-ci - Automate linting of shell-scripts with shellcheck.
 - jslint-ecma - Expand ES2015-feature-support for es-module-export-statement.
 - jslint - Relax warning expected_line_break_a_b for ternary-operator inside template-literal.
 - jslint - Add ability to auto-fix whitespace.
@@ -10,8 +11,9 @@
 - jslint - Try to improve parser to be able to parse jquery.js without stopping.
 
 # v2026.6.1-beta
-- jslint-ci - Rename ci-functions shGitPullrequestXxx to shGithubPrXxx.
 - jslint-ecma - Unify ES2015-destructure-logic into function prefix_destructure().
+- jslint-ci - Add shell-function shGithubPrUpdatePrxxx() to auto-update 'PR-xxx' placeholder to next sequential github issue/pull number.
+- jslint-ci - Rename shell-functions shGitPullrequestXxx() to shGithubPrXxx().
 - jslint-ecma - Update README.md, documenting supported ES2015+ features.
 - jslint - Update warning infix_in to recommend Object.hasOwn() over hasOwnProperty().
 - jslint-ecma - Add ES2025-feature RegExp Modifiers.